+++ CHANGES.md
@@ -7,6 +7,10 @@ Bugfixes
 - Fix `msvc_runtime_library` parameter in compile procedure of the
   `craftr.lang.cxx.msvc` module (invert behaviour of `'static'` and `'dynamic'`)
 
+Behaviour Changes
+
+- Explicit targets now also export a phony target rule in Ninja.
+
 API Changes
 
 - change `load()` now accepts a `mode` parameter instead of `get_namespace`,